% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.costexplorer_package.R
\docType{package}
\name{paws.costexplorer-package}
\alias{paws.costexplorer}
\alias{paws.costexplorer-package}
\title{paws.costexplorer: AWS Cost Explorer Service}
\description{
The Cost Explorer API allows you to
programmatically query your cost and usage data. You can query for
aggregated data such as total monthly costs or total daily usage. You
can also query for granular data, such as the number of daily write
operations for Amazon DynamoDB database tables in your production
environment.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
